requirements.txt
asn1crypto==0.24.0
beautifulsoup4==4.6.3
certifi==2018.10.15
cffi==1.11.5
cryptography==2.3.1
future==0.16.0
idna==2.7
pycparser==2.19
python-telegram-bot==11.1.0
six==1.11.0
